requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.3
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
bitsandbytes==0.49.1
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
databricks-sdk==0.82.0
datasets==4.5.0
dill==0.4.0
diskcache==5.6.3
docker==7.1.0
faiss-cpu==1.13.2
fastapi==0.128.0
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.3.5
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mlflow==3.9.0
mlflow-skinny==3.9.0
mlflow-tracing==3.9.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
narwhals==2.15.0
networkx==3.6.1
numpy==2.4.1
opentelemetry-api==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.6
packaging==25.0
pandas==2.3.3
peft==0.18.1
pillow==12.1.0
pluggy==1.6.0
prettytable==3.17.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
skops==0.13.0
smmap==5.0.2
SQLAlchemy==2.0.46
sqlparse==0.5.5
starlette==0.50.0
streamlit==1.53.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.1
transformers==5.0.0
trl==0.27.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.5.2
Werkzeug==3.1.5
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
